requirements.txt
chromadb==0.5.0
gradio==4.31.4
langchain==0.1.20
langchain-openai==0.1.6
pypdf==4.2.0
python-dotenv==1.0.1